Add -R to gpg_short_options for hidden recipients
Previous code displayed an getopt error that -R was an invalid
argument.  While -R was specified in long form, it was not
added to the gpg_short_options list.

Updated list to be in order (abcde instead of bacde).

This ordering was also updated in the gpg_allowed_options to
change from bacde to abcde.